requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
passlib[bcrypt]
pydantic[email]
pytest
httpx
